Improved @evaluate@ function readability & faster list parsing
Simplified the evaluate function by using 'maybe' and Applicatives
more aggressively. Case statements are ugly.

Improved the speed (and brevity) of proper list parsing by defining
a tail-recursive parser that accumulates a function to construct a
cons list that is passed 'Null' when there are no more items to parse.
This way, a list is not created using the slow 'many' Parsec combinator,
and that list is never traversed with 'foldr Cell Null'. Function
composition is arguably the most optimizeable operator, and tail recursion
is a reward unto itself!

Furthermore, the Parsec parser has been modified to work on ByteStrings.